+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This documents the main changes to the `tch` crate.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Improved the safetensor error wrapping,
+  [720](https://github.com/LaurentMazare/tch-rs/pull/720).
+
 ## v0.13.0 - 2023-05-18
 ### Added
 - Support static linking in the build script,